Fix crash when WebViewClient callback is called after fragment is done.
When the fragment is destroyed, the layout inflated by the fragment is
gone. This means we cannot access views (like progressBar) anymore. But
sometimes it happens that the WebViewClient invokes its callbacks after
the fragment has been destroyed. Since we access the progressBar view in
these callbacks, this leads to a crash.